Dev menu additions (#42)
* Карточку Projectile теперь можно прокачивать через Dev menu

* Magnet и CloneEveryone добавлены в дев-меню

* CollectAll и FastClaw были добавлены

* FreezeConveyor добавлена

* SlowDownBabushka добавлена

* SpawnRate добавлена

* WidenClaw добавлена

* Чуть изменил дев меню (убрал бесполезные кнопки)

+++ Assets/Resources/Effects/Projectile/Scripts/ProjectileSpawner.cs
@@ -4,7 +4,7 @@ namespace Resources.Effects.Projectile.Scripts
 {
     public class ProjectileSpawner : MonoBehaviour
     {
-        public float spawnInterval = 10f;
+        public int spawnInterval = 10;
         private float _previousSpawnInterval;
         public GameObject projectilePrefab;
